applications/supply_chain_demo/devices-simulation/requirements.txt
azure-iot-hub==2.6.1
azure-iot-device==2.12.0
